% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fffit.R
\name{anova.fffit}
\alias{anova.fffit}
\title{Comparing the Quality of the Possible Fits}
\usage{
\method{anova}{fffit}(object, ...)
}
\arguments{
\item{object}{An fffit object}

\item{...}{optional, additional arguments to anova. DON'T USE}
}
\value{
an anova object
}
\description{
\code{anova.fffit} prints an ANOVA table of all of the models that the fit made.
}
\seealso{
\verb{\link{anova}}, \verb{\link{nls}}
}
