% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits.R
\name{build.formula}
\alias{build.formula}
\title{A formula that describes a Finite Fourier Fit}
\usage{
build.formula(response, indices.to.use, all.phases)
}
\arguments{
\item{response}{A string or expression that will be the lhs. of the formula.}

\item{indices.to.use}{Which terms of the DFT to use in the formula.}

\item{all.phases}{Every phase from the DFT.
Elements are chosen with \code{indices.to.use}.}
}
\value{
a formula that represents a Finite Fourier Basis.
}
\description{
See \code{\link{build.terms}} for the details of how each term is built.
}
\examples{
build.formula("y - mu", 1:4, rnorm(10) \%\% pi)
}
