% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.R
\name{build.term.list}
\alias{build.term.list}
\title{A list of Fourier basis terms in the requested order}
\usage{
build.term.list(indices.to.use, all.phases)
}
\arguments{
\item{indices.to.use}{Which terms of the DFT to use in the formula.}

\item{all.phases}{Every phase from the DFT.
Elements are chosen with \code{indices.to.use}.}
}
\value{
a list of terms representing part of a Finite Fourier Basis.
}
\description{
See \code{\link{build.term}} for the details of how each term is built.
}
\examples{
build.term.list(1:4, rnorm(10) \%\% pi)
}
