% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits.R
\name{build.term}
\alias{build.term}
\title{One term from a Fourier basis}
\usage{
build.term(index, phase, variable.name = "a")
}
\arguments{
\item{index}{The integer that corresponds to an item in the DFT.}

\item{phase}{The actual phase of the item from the DFT.}

\item{variable.name}{An optional symbol that defaults to 'a'.}
}
\value{
A string representation of the term
}
\description{
A finite Fourier fit uses some of the terms from a Discrete Fourier
Transform to approximate a smooth fit to some data. This function
renders one such part of a term.
}
\examples{
build.term(1, 0)
build.term(42, 2.71828)
}
\seealso{
\code{\link{fft}} for calculating a DFT

\code{\link{Arg}} for calculating the phase of an item in DFT

\code{\link{as.formula}} for using the output to build a formula
}
