% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.argmin}
\alias{.argmin}
\title{The index of the minimum element in a container}
\usage{
.argmin(x)
}
\arguments{
\item{x}{Some collection, probably a \code{\link{vector}} of some kind.}
}
\value{
the index of the smallest item in the collection.
}
\description{
\code{argmin} acts just like the mathematical concept it is named after.
}
