% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fffit.R
\name{.build.model}
\alias{.build.model}
\title{A formula that describes a Finite Fourier Fit}
\usage{
.build.model(response, term.list, data, starts)
}
\arguments{
\item{response}{A string or expression that will be the lhs of the formula.}

\item{term.list}{The terms for the rhs of the formula.}

\item{data}{The data frame to predict on}

\item{starts}{The starting values for the \code{\link{nls}} fit.}
}
\value{
an \code{nls} object that represents a Finite Fourier Basis.
}
\description{
See \code{\link{build.term}} for the details of how each term is built.
}
