% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fffit.R
\name{fffit}
\alias{fffit}
\title{Fit a small, Fourier basis-like model to data}
\usage{
fffit(x, y, model.selector = BIC, max.terms = 10, pad.multiplier = 4)
}
\arguments{
\item{x}{Numeric, the independent variable}

\item{y}{Numeric, the dependent variable}

\item{model.selector}{Function, optional, the function for comparing models}

\item{max.terms}{Integer, optional, the greatest number of terms allowed}

\item{pad.multiplier}{Integer, optional, how much zero-padding to add}
}
\value{
an FFFit model
}
\description{
When a functional-looking relation is poorly approximated by the tools
\code{\link[=lm]{lm()}}, \code{\link[=glm]{glm()}}, or \code{\link[=nls]{nls()}}, one option is to use a linear combination of
cosine terms to approximate the relation within a strictly bounded domain.
This function does just that.
}
\examples{
x <- rnorm(10)
y <- 10*x + rnorm(10)
my.fit <- fffit(x, y)
}
