% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.R
\name{fffterm}
\alias{fffterm}
\title{Summarize a DFT Term for Use in an FFFit}
\usage{
fffterm(ranking, obj)
}
\arguments{
\item{ranking}{An integer representing where the term falls in some ordering}

\item{obj}{A \code{\link[=fourier.summary]{fourier.summary()}} object.
of the terms of the DFT.}
}
\value{
an object of type \code{fffterm}, which has the following named members:
a
~ The computed amplitude of the term

p
~ The phase of the term

f
~ The frequency of the term

term
~ A string with format $a\link{n}cos(2 pi f\link{n} x + p\link{n})$ for ranking \code{n}.
}
\description{
Several pieces of information from each term go in to building an \code{\link[=fffit]{fffit()}},
and it makes sense to calculate and store them once.
}
\seealso{
\code{\link[=fourier.summary]{fourier.summary()}}
}
