% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finitefourierfits.R
\docType{package}
\name{finitefourierfits}
\alias{finitefourierfits}
\title{Fit a Finite Fourier Basis to Data}
\description{
Sometimes there seems to be a clear functional relationship between two
variables in a data set, but it is poorly approximated by polynomials. A
Fourier basis may be able to help. It is, basically, a linear combination
of harmonically-related, phase-shifted, weighted cosine curves. This package
uses a very simple-minded approach to finding a basis with few curves.
}
