% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fffit.R
\name{formula.fffit}
\alias{formula.fffit}
\title{The Formula of the Best Fit}
\usage{
\method{formula}{fffit}(x, index = NULL, ...)
}
\arguments{
\item{x}{An fffit object}

\item{index}{optional, a specific model to get the formula for}

\item{...}{optional, additional arguments to \verb{\link{formula}}}
}
\value{
a formula
}
\description{
\code{formula.fffit} returns the formula of the best-fit model
}
\seealso{
\verb{\link{formula}}, \verb{\link{nls}}
}
