% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/half_complex.R
\name{fourier.transform}
\alias{fourier.transform}
\title{Compute a DFT, and Some Helpful Summaries and Descriptors}
\usage{
fourier.transform(x, sample.rate, multiplier = 4L)
}
\arguments{
\item{x}{A numeric vector to be Fourier transformed}

\item{sample.rate}{The sample rate of the time series underlying \code{x}.}

\item{multiplier}{optional, controls how much padding \code{x} gets.}
}
\value{
an S3 object of class \code{fourier.transform} with the following names:

dft
~ The complex-valued \code{\link[=dft]{dft()}} of x after it has been right-padded with zeros.

fft.size
~ The length of \code{dft}

sample.rate
~ This isn't actually used in the ctor, but it`s super useful to have here.

a
~ The real-valued magnitudes of the non-negative frequency term.

f
~ Frequencies in [0, Nyquist\)

p
~ The real-valued phases of each non-negative frequency term.

magnitude.order
~ Indices into \code{a} that sort it from largest to smallest.
}
\description{
\code{fourier.transform} zero-pads \code{x}, calculates the normal real to half-complex
DFT of the padded vector, and then stores the amplitudes and phases, and
amplitude rank of the non-negative frequency terms.
}
\seealso{
\code{\link[=dft]{dft()}}, \code{\link[=.amplitudes]{.amplitudes()}}, \code{\link[=.phases]{.phases()}}, \code{\link[=half.fft]{half.fft()}}
}
