% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier_summary.R
\name{half.fft}
\alias{half.fft}
\title{Half of the input, rounded down to the nearest integer.}
\usage{
half.fft(fft.size)
}
\arguments{
\item{fft.size}{The length of the Discrete Fourier Transform window.}
}
\value{
An integer equal to half of 'fft.size`, rounded down.
}
\description{
\code{half.fft} returns the pivot index between the positive and negative
frequencies in the output of \code{\link[=fft]{fft()}}.
}
