% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{omegas}
\alias{omegas}
\title{Angular frequencies for a Discrete Fourier Transform}
\usage{
omegas(fft.size)
}
\arguments{
\item{fft.size}{The size of the DFT}
}
\value{
a vector of floats of length \code{fft.size} in \eqn{[0, 2\pi)}.
}
\description{
If there are \code{fft.size} samples in the DFT then the frequencies run
from 0 (inclusive) to \eqn{2\pi} (exclusive).
}
\examples{
omegas(4)
}
