% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/half_complex.R
\name{padded.fft.size}
\alias{padded.fft.size}
\title{The Target Fft Size for an \verb{\link{fffit}} Object Based on Input Size}
\usage{
padded.fft.size(x, multiplier = 4L)
}
\arguments{
\item{x}{An array, vector, or integer, which will be used to calculate the
input length using \code{dim}, \code{length}, or the integer itself, respectively.}

\item{multiplier}{optional, an integer multiplier >= 1. defaults to 4}
}
\value{
The length of the input, rounded up to the nearest power of two and
multiplied by \code{multiplier}.
}
\description{
Rounding up to the nearest power of two is traditional for computing the
length of a DFT
}
\seealso{
\verb{\link{nextn}}
}
