% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fffit.R
\name{predict.fffit}
\alias{predict.fffit}
\title{Predicting from Finite Fourier Fits}
\usage{
\method{predict}{fffit}(object, newdata = NULL, index = NULL, ...)
}
\arguments{
\item{object}{An \code{fffit} object}

\item{newdata}{A data frame with new independent and dependent variables.}

\item{index}{optional, a specific model to summarize}

\item{...}{Additional arguments to \code{predict.nls}}
}
\value{
A numeric vector of predictions
}
\description{
\code{predict.fffit} produces predicted values, obtained by evaluating the fit
in the context of the \code{newdata}. This will give nonsensical results if the
input domain is not within the bounds of the original analysis's domain.
}
\seealso{
\code{\link{predict.nls}}
}
