% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fffit.R
\name{coef.fffit}
\alias{coef.fffit}
\title{The Coefficients of the Best Fit}
\usage{
\method{coef}{fffit}(object, index = NULL, ...)
}
\arguments{
\item{object}{An fffit object}

\item{index}{optional, a specific model to extract coefficients from.}

\item{...}{optional, additional arguments to \code{\link[=coef]{coef()}}}
}
\value{
a vector of coefficients
}
\description{
\code{coef.fffit} returns the coefficients of the best-fit model
}
\seealso{
\code{\link[=coef]{coef()}}, \code{\link[=nls]{nls()}}
}
