% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.R
\name{fffterm}
\alias{fffterm}
\title{Summarize a DFT Term for Use in an FFFit}
\usage{
fffterm(ranking, obj)
}
\arguments{
\item{ranking}{An integer representing where the term falls in some ordering
of the terms of the DFT.}

\item{obj}{A \code{\link[=fourier.summary]{fourier.summary()}} object.}
}
\value{
an object of type \code{fffterm}, which has the following named members:
\describe{
\item{a}{The computed amplitude of the term}
\item{p}{The phase of the term}
\item{f}{The frequency of the term}
\item{term}
{A string,
"\eqn{a\mathrm{n}\,\cos(2\pi\,f\mathrm{n}\,x + p\mathrm{n}})",
for the ranking \code{n}.}
}
}
\description{
Several pieces of information from each term go in to building an \code{\link[=fffit]{fffit()}},
and it makes sense to calculate and store them once.
}
\seealso{
\code{\link[=fourier.summary]{fourier.summary()}}
}
