% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fffit.R
\name{predict.fffit}
\alias{predict.fffit}
\title{Predicting from Finite Fourier Fits}
\usage{
\method{predict}{fffit}(fit, newdata, ...)
}
\arguments{
\item{fit}{An \code{fffit} object}

\item{newdata}{A new independent variable.}
}
\value{
A numeric vector of predictions
}
\description{
\code{predict.fffit} produces predicted values, obtained by evaluating the fit
in the context of the \code{newdata}. This will give nonsensical results if the
input domain is not within the bounds of the original analysis's domain.
}
\seealso{
\code{\link{predict.nls}}
}
