% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fffit.R
\name{summary.fffit}
\alias{summary.fffit}
\title{Demonstrating the Quality of the Best Fit}
\usage{
\method{summary}{fffit}(object, index = NULL, ...)
}
\arguments{
\item{object}{An fffit object}

\item{index}{optional, a specific model to summarize}

\item{...}{optional, additional arguments to \verb{\link{summary}}}
}
\value{
a summary object
}
\description{
\code{summary.fffit} prints an summary table of the model that fit best.
}
\seealso{
\verb{\link{summary}}, \verb{\link{nls}}
}
